% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_cosine_similarity}
\alias{calculate_cosine_similarity}
\title{calculate_cosine_similarity}
\usage{
calculate_cosine_similarity(a, b)
}
\arguments{
\item{b}{}
}
\description{
Calculate cosine similarity of two vectors. Function adapted from the
MicroRaman package.
}
