% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_spectrum.R
\name{make_every_spectrum}
\alias{make_every_spectrum}
\title{make_every_spectrum}
\usage{
make_every_spectrum(
  make_every_XIC_output,
  mz_window = 3,
  SN_cutoff = 10,
  mean_cosine_sim_cutoff = 0.8,
  mean_score_mfa_cutoff = 0.3,
  resPowerMS1 = 3e+05,
  isotopologue_window_multiplier = 6,
  MSoutputWidth = 18,
  MSoutputDPI = 200,
  makePNG = FALSE,
  return_timers = TRUE,
  save_spec_object = FALSE
)
}
\arguments{
\item{make_every_XIC_output}{Direct output from make_every_XIC.}

\item{mz_window}{Width of the m/z window to be used to make mini spectra.}

\item{MSoutputWidth}{Width of png output. Only used is makePNG is TRUE.}

\item{MSoutputDPI}{DPI of png output. Only used is makePNG is TRUE.}

\item{makePNG}{Whether to make png output. Probably deprecated soon.}

\item{return_timers}{Controls whether timers or spectra are returned.}
}
\value{

}
\description{
make_every_spectrum
}
