% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_XIC_MS1.R
\name{make_every_XIC_MS1}
\alias{make_every_XIC_MS1}
\title{make_every_XIC_MS1}
\usage{
make_every_XIC_MS1(
  rawFileDir = NULL,
  rawFileName = NULL,
  targetSeqData = NULL,
  outputDir = getwd(),
  target_col_name = c("UNIPROTKB"),
  target_sequence_col_name = c("ProteoformSequence"),
  PTMname_col_name = c("PTMname"),
  PTMformula_col_name1 = c("FormulaToAdd"),
  PTMformula_col_name2 = c("FormulaToSubtract"),
  isoAbund = c(`12C` = 0.9893, `14N` = 0.99636),
  target_charges = c(1:50),
  mass_range = c(0, 1e+05),
  mz_range = c(600, 2000),
  abund_cutoff = 15,
  sample_n_pforms = NULL,
  XIC_tol = 2,
  use_IAA = FALSE,
  save_output = TRUE,
  mean_score_mfa_cutoff = 0.3,
  mean_cosine_sim_cutoff = 0.8,
  SN_cutoff = 10,
  resPowerMS1 = 3e+05,
  isotopologue_window_multiplier = 6,
  mz_window = 3,
  return_timers = TRUE,
  save_spec_object = FALSE,
  rawrrTemp = tempdir()
)
}
\arguments{
\item{abund_cutoff}{}
}
\value{

}
\description{
make_every_XIC_MS1
}
