% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_XIC_MS2.R
\name{make_every_XIC_MS2}
\alias{make_every_XIC_MS2}
\title{make_every_XIC_MS2}
\usage{
make_every_XIC_MS2(
  rawFileDir = NULL,
  rawFileName = NULL,
  targetSeqData = NULL,
  outputDir = getwd(),
  target_col_name = "UNIPROTKB",
  target_sequence_col_name = "ProteoformSequence",
  PTMname_col_name = "PTMname",
  PTMformula_col_name1 = "FormulaToAdd",
  PTMformula_col_name2 = "FormulaToSubtract",
  isoAbund = c(`12C` = 0.9893, `14N` = 0.99636),
  fragment_charges = c(1:50),
  fragment_types = c("b", "y"),
  fragment_mz_range = c(300, 2000),
  fragment_pos_cutoff = c(1, 50),
  XIC_tol_MS2 = 10,
  XIC_cutoff = 5e-04,
  scoreMFAcutoff = 0.3,
  cosinesimcutoff = 0.99,
  resPowerMS2 = 150000,
  isotopologue_window_multiplier = 6,
  mz_window = 5,
  use_IAA = FALSE,
  save_output = TRUE,
  abund_cutoff = 5
)
}
\arguments{
\item{abund_cutoff}{}
}
\value{

}
\description{
make_every_XIC_MS2
}
