% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_spectrum.R
\name{make_every_spectrum}
\alias{make_every_spectrum}
\title{make_every_spectrum}
\usage{
make_every_spectrum(
  make_every_XIC_output,
  mz_window = 3,
  mz_window_scaling = 0.00333,
  SN_cutoff = 10,
  mean_cosine_sim_cutoff = 0.8,
  MSoutputWidth = 18,
  MSoutputDPI = 200,
  makePNG = FALSE,
  future_workers = 1L,
  return_timers = TRUE
)
}
\arguments{
\item{make_every_XIC_output}{Direct output from make_every_XIC.}

\item{mz_window}{Width of the m/z window to be used to make mini spectra.}

\item{mz_window_scaling}{Scaling factor for window for determination of intensity
for individual isotopologue peaks. Multiplied by mz_window value.}

\item{MSoutputWidth}{Width of png output. Only used is makePNG is TRUE.}

\item{MSoutputDPI}{DPI of png output. Only used is makePNG is TRUE.}

\item{makePNG}{Whether to make png output. Probably deprecated soon.}

\item{return_timers}{Controls whether timers or spectra are returned.}
}
\value{

}
\description{
make_every_spectrum
}
