% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_XIC.R
\name{make_every_XIC}
\alias{make_every_XIC}
\title{make_every_XIC}
\usage{
make_every_XIC(
  rawFileDir = NULL,
  rawFileName = NULL,
  targetSeqData = NULL,
  outputDir = getwd(),
  target_col_name = c("UNIPROTKB", "PFR"),
  target_sequence_col_name = c("ProteoformSequence"),
  PTMname_col_name = c("PTMname"),
  PTMformula_col_name1 = c("FormulaToAdd"),
  PTMformula_col_name2 = c("FormulaToSubtract"),
  use_depleted_isotopes = FALSE,
  top_n_pforms = NULL,
  mass_range = c(0, 50000),
  target_charges = c(1:30),
  mz_range = c(600, 2000),
  XIC_tol = 25,
  use_IAA = FALSE,
  abund_cutoff = 5
)
}
\arguments{
\item{abund_cutoff}{}
}
\value{

}
\description{
make_every_XIC
}
