% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_XIC.R
\name{make_every_XIC}
\alias{make_every_XIC}
\title{make_every_XIC}
\usage{
make_every_XIC(
  rawFileDir = NULL,
  rawFileName = NULL,
  targetSeqData = NULL,
  outputDir = getwd(),
  target_col_name = c("UNIPROTKB"),
  target_sequence_col_name = c("ProteoformSequence"),
  PTMname_col_name = c("PTMname"),
  PTMformula_col_name1 = c("FormulaToAdd"),
  PTMformula_col_name2 = c("FormulaToSubtract"),
  use_depleted_isotopes = FALSE,
  sample_n_pforms = NULL,
  mass_range = c(0, 1e+05),
  target_charges = c(1:50),
  mz_range = c(600, 2000),
  XIC_tol = 2,
  use_IAA = FALSE,
  save_output = TRUE,
  abund_cutoff = 5,
  future_workers = 8
)
}
\arguments{
\item{abund_cutoff}{}
}
\value{

}
\description{
make_every_XIC
}
