% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_every_XIC_MS1.R
\name{make_every_XIC_MS1}
\alias{make_every_XIC_MS1}
\title{make_every_XIC_MS1}
\usage{
make_every_XIC_MS1(
  rawFileDir = NULL,
  rawFileName = NULL,
  targetSeqData = NULL,
  outputDir = getwd(),
  target_col_name = "UNIPROTKB",
  target_sequence_col_name = "ProteoformSequence",
  PTMname_col_name = "PTMname",
  PTMformula_col_name1 = "FormulaToAdd",
  PTMformula_col_name2 = "FormulaToSubtract",
  isoAbund = c(`12C` = 0.9893, `14N` = 0.99636),
  target_charges = c(1:50),
  mass_range = c(0, 1e+05),
  mz_range = c(600, 2000),
  abund_cutoff = 15,
  sample_n_pforms = NULL,
  XIC_tol = 2,
  use_IAA = FALSE,
  save_output = TRUE,
  scoreMFAcutoff = 0.3,
  cosinesimcutoff = 0.8,
  SN_cutoff = 10,
  resPowerMS1 = 3e+05,
  isotopologue_window_multiplier = 6,
  mz_window = 3,
  return_timers = TRUE,
  rawrrTemp = tempdir(),
  save_spec_object = FALSE
)
}
\arguments{
\item{save_spec_object}{}
}
\value{

}
\description{
make_every_XIC_MS1
}
